% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{renderer}
\alias{renderer}
\alias{rmarkdown_renderer}
\alias{rmdmatter_renderer}
\alias{renaming_renderer}
\title{Create a function to render R Markdown with pre-filled options}
\usage{
rmarkdown_renderer(...)

rmdmatter_renderer(...)

renaming_renderer(callback, ...)
}
\arguments{
\item{...}{Arguments passed to the underlying rendering function when the
returned function is called.}

\item{callback}{A function that gets the rendered R Markdown YAML front
matter metadata as an argument, and returns a new name for \code{file}. Any
changes implied to the directory or file extension are ignored -- only the
name is changed.}
}
\value{
A function that takes two arguments, \code{input} and \code{encoding}, and
renders the \code{input} R Markdown document when called.
}
\description{
The returned function is compatible with the \code{knit} YAML front matter field
in R Markdown documents, which lets you \href{https://bookdown.org/yihui/rmarkdown-cookbook/custom-knit.html}{customize the \code{Knit} button in RStudio}.
}
\details{
Renderer functions are particularly useful in conjunction with the \code{knit}
metadata field to specify custom rendering options in RStudio. For example,
extending the example header in \link[rmarkdown:rmd_metadata]{rmarkdown::rmd_metadata}, you can easily
specify a fixed output file with \code{rmarkdown_renderer()}:\preformatted{---
title: "Crop Analysis Q3 2013"
author: Martha Smith
date: October 23rd, 2013
knit: rmdmatter::rmarkdown_renderer(output_file = "analysis-q3")
---
}

Or use \code{renaming_renderer()} to name the output dynamically using fields in
the front matter:\preformatted{---
title: "Crop Analysis Q3 2013"
author: Martha Smith
date: October 23rd, 2013
knit: |
  rmdmatter::renaming_renderer(function(metadata) metadata$title)
---
}
}
\section{Functions}{
\itemize{
\item \code{rmarkdown_renderer}: Vanilla R Markdown rendering.

\item \code{rmdmatter_renderer}: Retains the rendered YAML front matter.

\item \code{renaming_renderer}: Retains the rendered YAML front matter, and passes it to
\code{callback} to rename the output file.
}}

\seealso{
\code{\link[rmarkdown:render]{rmarkdown::render()}} used by \code{rmarkdown_renderer()}.

\code{\link[=render]{render()}} used by \code{rmdmatter_renderer()}.

\code{\link[=render_and_rename]{render_and_rename()}} used by \code{renaming_renderer()}.
}
